Added right click looping Also refactored to use BlockPos instead of x,y,z

+++ src/main/java/com/zoonie/InteractionSounds/sound/Sound.java
@@ -13,7 +13,7 @@ public class Sound implements Serializable
 	private SoundState state;
 	@Expose
 	private float volume;
-	private double length;
+	private long length;
 	private long size;
 
 	public Sound(Sound sound)
@@ -83,7 +83,7 @@ public class Sound implements Serializable
 		this.volume = volume;
 	}
 
-	public Double getLength()
+	public long getLength()
 	{
 		return length;
 	}
