+++ src/main/java/com/zoonie/InteractionSounds/sound/SoundInfo.java
@@ -3,7 +3,7 @@ package com.zoonie.InteractionSounds.sound;
 public class SoundInfo implements Comparable
 {
 	public String name, category;
-	public double length;
+	public long length;
 	public long size;
 
 	public SoundInfo(String name, String category)
@@ -12,7 +12,7 @@ public class SoundInfo implements Comparable
 		this.category = category;
 	}
 
-	public SoundInfo(String name, String category, double length, long size)
+	public SoundInfo(String name, String category, long length, long size)
 	{
 		this.name = name;
 		this.category = category;
